Extract shm logic from POSIX
