+++ qiling/os/posix/posix.py
@@ -4,7 +4,7 @@
 #
 
 from inspect import signature, Parameter
-from typing import Dict, TextIO, Tuple, Union, Callable, IO, List, Optional
+from typing import Dict, TextIO, Union, Callable, IO, List, Optional
 
 from unicorn.arm64_const import UC_ARM64_REG_X8, UC_ARM64_REG_X16
 from unicorn.arm_const import (
@@ -26,7 +26,9 @@ from qiling.cc import QlCC, intel, arm, mips, riscv, ppc
 from qiling.const import QL_ARCH, QL_OS, QL_INTERCEPT
 from qiling.exception import QlErrorSyscallNotFound
 from qiling.os.os import QlOs
+from qiling.os.posix.const import NR_OPEN, errors
 from qiling.os.posix.msq import QlMsq
+from qiling.os.posix.shm import QlShm
 from qiling.utils import ql_get_module, ql_get_module_function
 
 SYSCALL_PREF: str = f'ql_syscall_'
@@ -89,51 +91,6 @@ class QlFileDes:
 
     def restore(self, fds):
         self.__fds = fds
-
-
-# vaguely reflects a shmid64_ds structure
-class QlShmId:
-
-    def __init__(self, key: int, uid: int, gid: int, mode: int, segsz: int) -> None:
-        # ipc64_perm
-        self.key = key
-        self.uid = uid
-        self.gid = gid
-        self.mode = mode
-
-        self.segsz = segsz
-
-        # track the memory locations this segment is currently attached to
-        self.attach: List[int] = []
-
-
-class QlShm:
-    def __init__(self) -> None:
-        self.__shm: Dict[int, QlShmId] = {}
-        self.__id: int = 0x0F000000
-
-    def __len__(self) -> int:
-        return len(self.__shm)
-
-    def add(self, shm: QlShmId) -> int:
-        shmid = self.__id
-        self.__shm[shmid] = shm
-
-        self.__id += 0x1000
-
-        return shmid
-
-    def remove(self, shmid: int) -> None:
-        del self.__shm[shmid]
-
-    def get_by_key(self, key: int) -> Tuple[int, Optional[QlShmId]]:
-        return next(((shmid, shmobj) for shmid, shmobj in self.__shm.items() if shmobj.key == key), (-1, None))
-
-    def get_by_id(self, shmid: int) -> Optional[QlShmId]:
-        return self.__shm.get(shmid, None)
-
-    def get_by_attaddr(self, shmaddr: int) -> Optional[QlShmId]:
-        return next((shmobj for shmobj in self.__shm.values() if shmobj.attach.count(shmaddr) > 0), None)
 
 
 class QlOsPosix(QlOs):
